scripts/requirements.txt
# Aibō Voice Service - Python Dependencies
# Install with: pip install -r requirements.txt

deepgram-sdk==3.0.0
PyAudio>=0.2.13
python-dotenv>=1.0.0
